Create IFleetRepository.cs
The IFleetRepository interface in the OZFleet.Core.Interfaces namespace defines the contract for the repository responsible for managing Fleet entities in the system. It contains the following methods that will be implemented by a concrete repository class to handle the interactions with the data store (e.g., database, file system, etc.).